Removed all bash documentation
