--- docs/html/referenceTemplate.html
@@ -1,63 +1,0 @@
-<!doctype html>
-
-<html lang="en">
-<head>
-  <meta charset="utf-8">
-  <title>Hail</title>
-  <meta name="description" content="Hail">
-  <link rel='shortcut icon' href='hail_logo_sq.png' type='image/x-icon'>
-  <script  src="jquery-3.1.1.min.js"></script>
-  <script src="bootstrap.min.js"></script>
-  <link rel="stylesheet" href="bootstrap.min.css" type="text/css">
-  <link rel="stylesheet" href="style.css">
-  <link rel="stylesheet" href="navbar.css">
-  <script>
-    $(document).ready(function () {
-      $("#hail-navbar").load("navbar.html", function () {
-        $(".nav li").removeClass("active");
-        $('#docs').addClass('active');
-        $('#reference').addClass('active');
-      });
-    });
-  </script>
-</head>
-
-<body>
-  <nav class="navbar navbar-default navbar-static-top" id="hail-navbar"></nav>
-  <div id="body">
-  <div id="TOC">
-    <h1>General Reference</h1>
-    <h2>Table of Contents</h2>
-    <ul>
-      <li><a href="#Representation">Data Representation</a></li>
-      <li><a href="#Importing">Importing Genotype Data</a></li>
-      <li><a href="#HailObjectProperties">Hail Object Methods</a></li>
-      <ul id="toc-hailobjects">
-        <li><a href="#variant">Variant</a></li>
-        <li><a href="#altallele">AltAllele</a></li>
-        <li><a href="#locus">Locus</a></li>
-        <li><a href="#interval">Interval</a></li>
-        <li><a href="#sample">Sample</a></li>
-        <li><a href="#genotype">Genotype</a></li>
-      </ul>
-      <li><a href="#Annotations">Annotations</a></li>
-      <li><a href="#HailExpressionLanguage">Hail Expression Language</a></li>
-      <li><a href="#Filtering">Filtering Data</a></li>
-      <li><a href="#ExportingData">Exporting Genotype Data</a></li>
-      <li><a href="#ExportingTSV">Exporting Data to TSV</a></li>
-      <li><a href="#SQL">SQL</a></li>
-    </ul>
-  </div>
-
-<div id="Representation"></div>
-<div id="Importing"></div>
-<div id="HailObjectProperties"></div>
-<div id="Annotations"></div>
-<div id="HailExpressionLanguage"></div>
-<div id="Filtering"></div>
-<div id="ExportingData"></div>
-<div id="ExportingTSV"></div>
-<div id="SQL"></div>
-  </div>
-</body>
-</html>
